move more code to propertyReader
